github_python_slack/requirements.txt
beautifulsoup4
requests
line-bot-sdk